IMX6SL: Register clock for UART5
Clock for UART5 needs to be registered to be used.
IMX6Q has a conflict between UART4/5 and PCIE, but
it is not clear to me why UART5 clock is not registered
by default. This note is to start a trail if this ends
up uncovering an undocumented conflict with UART5.

Signed-off-by: Jacob Postman <jacob@revolution-robotics.com>